refactor: move user controller to its own folder
