+++ src/controllers/user/delete-user.js
@@ -4,7 +4,7 @@ import {
     invalidIdResponse,
     ok,
     userNotFoundResponse,
-} from './helpers/index.js'
+} from '../helpers/index.js'
 
 export class DeleteUserController {
     constructor(deleteUserUseCase) {
